extract package, fix build
