+++ jest.config.js
@@ -5,5 +5,6 @@ module.exports = {
   modulePaths: ["<rootDir>"],
   moduleNameMapper: {
     "reactive-crdt": "<rootDir>/packages/reactive-crdt/src",
+    "yjs-reactive-bindings": "<rootDir>/packages/yjs-reactive-bindings/src",
   },
 };
