+++ packages/yjs-reactive-bindings/.npmignore
@@ -1,0 +1,1 @@
+test
