+++ packages/yjs-reactive-bindings/src/observableProvider.ts
