+++ packages/yjs-reactive-bindings/src/types/map.ts
@@ -1,4 +1,4 @@
-import { Atom, createAtom } from "./observableProvider";
+import { Atom, createAtom } from "../observableProvider";
 import * as Y from "yjs";
 
 const mapsObserved = new WeakSet<Y.Map<any>>();
